+++ assembly-syntax/src/ast/instruction/mod.rs
@@ -3,6 +3,7 @@ pub mod debug;
 mod print;
 
 use alloc::vec::Vec;
+use core::ops::Range;
 
 pub use self::{advice::SystemEventNode, debug::DebugOptions};
 use crate::{
@@ -211,9 +212,7 @@ pub enum Instruction {
     PushU32(u32),
     PushFelt(Felt),
     PushWord(WordValue),
-    PushU8List(Vec<u8>),
-    PushU16List(Vec<u16>),
-    PushU32List(Vec<u32>),
+    PushSlice(Immediate<IntValue>, Range<usize>),
     PushFeltList(Vec<Felt>),
     Locaddr(ImmU16),
     Sdepth,